Fixed up code to use size_t in preference to int - removed redundant casts and changes to use npos instead of -1.

+++ ConnectDlg.hpp
@@ -36,7 +36,7 @@ public:
 	//
 	// Members.
 	//
-	int		m_nConnection;
+	size_t	m_nConnection;
 	CString	m_strLogin;
 	CString	m_strPassword;
